+++ src/generate/r3f/utils.ts
@@ -1,7 +1,12 @@
 import { Object3D } from 'three'
 
-import { AnalyzedGLTF } from '../analyze/AnalyzedGLTF.js'
-import { isBone, isInstancedMesh, isMesh, isTargetedLight } from '../analyze/is.js'
+import {
+  AnalyzedGLTF,
+  isBone,
+  isInstancedMesh,
+  isMesh,
+  isTargetedLight,
+} from '../../analyze/index.js'
 
 /**
  * r3f specific determination of jsx component?  If not, this could be moved to AnalyzeGLTF.
