+++ src/index.ts
@@ -1,16 +1,6 @@
-export * from './generate/AbstractGenerate.js'
-export * from './analyze/AnalyzedGLTF.js'
-export * from './analyze/is.js'
-export * from './analyze/isVarName.js'
-export * from './analyze/pruneStrategies.js'
-export * from './analyze/shallowEqual.js'
-export * from './analyze/utils.js'
-export * from './loadGLTF.js'
-export * from './Log.js'
+export * from './analyze/index.js'
+export * from './generate/index.js'
+export * from './load/index.js'
 export * from './options.js'
-export * from './r3f/GenerateR3F.js'
-export * from './r3f/utils.js'
-export * from './transform/gltfTransform.js'
-export * from './transform/utils.js'
-export * from './utils/files.js'
-export * from './utils/types.js'
+export * from './transform/index.js'
+export * from './utils/index.js'
